+++ index.js
@@ -9,10 +9,10 @@ require('dotenv').config();
 const cookieParser = require('cookie-parser');
 
 // Production DB connection string
-const dbConnectionString = process.env.MONGO_URI;
+// const dbConnectionString = process.env.MONGO_URI;
 
 // Development DB connection string
-// const dbConnectionString = 'mongodb://127.0.0.1:27017/Race-Fanatic-DB';
+const dbConnectionString = 'mongodb://127.0.0.1:27017/Race-Fanatic-DB';
 
 const initializeDatabase = () =>
   mongoose
